+++ example/lib/pages/home_page.dart
@@ -34,7 +34,9 @@ class _HomePageState extends State<HomePage> {
 
   Future<void> _loadFromAssets() async {
     try {
-      final result = await rootBundle.loadString(isDesktop() ? 'assets/sample_data_nomedia.json' : 'assets/sample_data.json');
+      final result = await rootBundle.loadString(isDesktop()
+          ? 'assets/sample_data_nomedia.json'
+          : 'assets/sample_data.json');
       final doc = Document.fromJson(jsonDecode(result));
       setState(() {
         _controller = QuillController(