+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+# [6.0.1]
+* Changed translation country code (zh_HK -> zh_hk) to lower case, which is required for i18n_extension used in flutter_quill.
+* Add localization in example's main to demonstrate translation.
+* Issue [Windows] selection's copy / paste tool bar not shown #861: add selection's copy / paste toolbar, escape to hide toolbar, mouse right click to show toolbar, ctrl-Y / ctrl-Z to undo / redo.
+* Image and video displayed in Windows platform caused screen flickering while selecting text, a sample_data_nomedia.json asset is added for Desktop to demonstrate the added features.
+* Known issue: keyboard action sometimes causes exception mentioned in Flutter's issue #106475 ([Windows] Keyboard shortcuts stop working after modifier key repeat flutter/flutter#106475).
+* Know issue: user needs to click the editor to get focus before toolbar is able to display.
+
 # [6.0.0] BREAKING CHANGE
-* Removed embed (image, video & forumla) blocks from the package to reduce app size.
+* Removed embed (image, video & formula) blocks from the package to reduce app size.
 
 These blocks have been moved to the package `flutter_quill_extensions`, migrate by filling the `embedBuilders` and `embedButtons` parameters as follows: